// src/app/models/admin/dangkyhoatdongkhoangsan/dkkhaithaccongtrinh.model.ts
export class OutputDkKhaiThacCongTrinh {
  idcongtrinh: string;
  sohieu: string;
  chieusau: number;
  toadox: number;
  toadoy: number;
  iddangkykhaithac: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  ghichu: string;
  siteid: string;
}

export class InputDkKhaiThacCongTrinh {
  idcongtrinh: string;
  sohieu: string;
  chieusau: number;
  toadox: number;
  toadoy: number;
  iddangkykhaithac: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  ghichu: string;
  siteid: string;
}